// src/components/blocks/types/Html.ts
export enum ElementType {
  BUTTON = "BUTTON", // <button></button>
  ANCHOR = "ANCHOR", // <a></a>
  IMAGE = "IMAGE", // <img/>
  PARAGRAPH = "PARAGRAPH", // <p></p>
  DIVISION = "DIVISION", // <div></div>
  SPAN = "SPAN", // <span></span>
  LIST_ITEM = "LIST_ITEM", // <span></span>
  INPUT_FIELD = "INPUT_FIELD", // <input/>
  HEADING = "HEADING", // <h1></h1> , ...
  TEXT_AREA = "TEXT_AREA", //<textarea></textarea>
  ICON = "ICON", //<i></id>
  ICONS_ARRAY = "ICONS_ARRAY",
  ITEMS_ARRAY = "ITEMS_ARRAY",
  VIDEO = "VIDEO",
}

export enum ElementPosition {
  START = "START",
  END = "END",
}
export enum FlexDirection {
  COLUMN = "COLUMN",
  ROW = "ROW",
}

export enum ResponsiveVisibility {
  Mobile = "Mobile",
  Desktop = "Desktop",
}